requirements.txt
beautifulsoup4==4.8.1
dj-database-url==0.5.0
Django==2.1.5
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-taggit==0.22.2
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.7.5
pytz==2018.5
soupsieve==1.9.4
whitenoise==4.0
openpyxl==2.6.0
xlwt
djangorestframework==3.10.3

